refactor: modificate init app
