+++ src/app.py
@@ -2,7 +2,6 @@ from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
 from datetime import timedelta
 from flask_login import LoginManager
-import os
 
 app = Flask(__name__, instance_relative_config=True)
 
@@ -12,6 +11,11 @@ login_manager.init_app(app)
 app.config.from_object("config.Config")
 
 db = SQLAlchemy(app)
+
+with app.app_context():
+    from src.routes import handle
+
+    app.register_blueprint(handle)
 
 app.permanent_session_lifetime = timedelta(
     days=int(app.config["DAYS"])
